move schemas into their own folder - there will be more of them
