+++ flexmeasures/api/common/schemas/times.py
